031802518/requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.14.61
botocore==1.17.61
certifi==2020.6.20
chardet==3.0.4
coverage==5.3
Cython==0.29.14
Django==3.1.2
docutils==0.15.2
gensim==3.8.3
idna==2.10
image==1.5.32
jieba==0.42.1
jmespath==0.10.0
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.2
six==1.15.0
smart-open==2.1.1
sqlparse==0.3.1
tk==0.1.0
urllib3==1.25.10
